+++ cli/src/commands/util/install_man_pages.rs
@@ -12,24 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::io::Write as _;
+use std::path::PathBuf;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
 use crate::ui::Ui;
 
-/// Print a ROFF (manpage)
+/// Install Jujutsu's manpages to the provided path
 #[derive(clap::Args, Clone, Debug)]
-pub struct UtilMangenArgs {}
+pub struct UtilInstallManPagesArgs {
+    /// The path where manpages will installed. An example path might be
+    /// `/usr/share/man`. The provided path will be appended with `man1`,
+    /// etc., as appropriate
+    path: PathBuf,
+}
 
-pub fn cmd_util_mangen(
-    ui: &mut Ui,
+pub fn cmd_util_install_man_pages(
+    _ui: &mut Ui,
     command: &CommandHelper,
-    _args: &UtilMangenArgs,
+    args: &UtilInstallManPagesArgs,
 ) -> Result<(), CommandError> {
-    let mut buf = vec![];
-    let man = clap_mangen::Man::new(command.app().clone());
-    man.render(&mut buf)?;
-    ui.stdout().write_all(&buf)?;
+    let man1_dir = args.path.join("man1");
+    std::fs::create_dir_all(&man1_dir)?;
+    let app = command.app().clone();
+    clap_mangen::generate_to(app, man1_dir)?;
     Ok(())
 }