+++ cli/src/commands/util/mod.rs
@@ -16,7 +16,7 @@ mod completion;
 mod config_schema;
 mod exec;
 mod gc;
-mod mangen;
+mod install_man_pages;
 mod markdown_help;
 
 use clap::Subcommand;
@@ -30,8 +30,8 @@ use self::exec::cmd_util_exec;
 use self::exec::UtilExecArgs;
 use self::gc::cmd_util_gc;
 use self::gc::UtilGcArgs;
-use self::mangen::cmd_util_mangen;
-use self::mangen::UtilMangenArgs;
+use self::install_man_pages::cmd_util_install_man_pages;
+use self::install_man_pages::UtilInstallManPagesArgs;
 use self::markdown_help::cmd_util_markdown_help;
 use self::markdown_help::UtilMarkdownHelp;
 use crate::cli_util::CommandHelper;
@@ -45,7 +45,7 @@ pub(crate) enum UtilCommand {
     ConfigSchema(UtilConfigSchemaArgs),
     Exec(UtilExecArgs),
     Gc(UtilGcArgs),
-    Mangen(UtilMangenArgs),
+    InstallManPages(UtilInstallManPagesArgs),
     MarkdownHelp(UtilMarkdownHelp),
 }
 
@@ -60,7 +60,7 @@ pub(crate) fn cmd_util(
         UtilCommand::ConfigSchema(args) => cmd_util_config_schema(ui, command, args),
         UtilCommand::Exec(args) => cmd_util_exec(ui, command, args),
         UtilCommand::Gc(args) => cmd_util_gc(ui, command, args),
-        UtilCommand::Mangen(args) => cmd_util_mangen(ui, command, args),
+        UtilCommand::InstallManPages(args) => cmd_util_install_man_pages(ui, command, args),
         UtilCommand::MarkdownHelp(args) => cmd_util_markdown_help(ui, command, args),
     }
 }